refactor llm sessions
