+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/agent/session/AIAgentLLMWriteSession.kt
@@ -10,22 +10,10 @@ import ai.koog.agents.core.environment.SafeTool
 import ai.koog.agents.core.tools.Tool
 import ai.koog.agents.core.tools.ToolDescriptor
 import ai.koog.agents.core.tools.ToolRegistry
-import ai.koog.prompt.dsl.ModerationResult
 import ai.koog.prompt.dsl.Prompt
-import ai.koog.prompt.dsl.PromptBuilder
 import ai.koog.prompt.executor.model.PromptExecutor
-import ai.koog.prompt.executor.model.StructureFixingParser
 import ai.koog.prompt.llm.LLModel
-import ai.koog.prompt.message.LLMChoice
-import ai.koog.prompt.message.Message
-import ai.koog.prompt.params.LLMParams
 import ai.koog.prompt.processor.ResponseProcessor
-import ai.koog.prompt.streaming.StreamFrame
-import ai.koog.prompt.structure.StructureDefinition
-import ai.koog.prompt.structure.StructuredRequestConfig
-import ai.koog.prompt.structure.StructuredResponse
-import kotlinx.coroutines.flow.Flow
-import kotlinx.serialization.KSerializer
 import kotlin.reflect.KClass
 import kotlin.time.Clock
 
@@ -34,18 +22,22 @@ import kotlin.time.Clock
  * and tools in an agent environment. This class provides functionality for executing
  * LLM requests, managing tools, and customizing prompts dynamically within a specific
  * session context.
- *
- * @param environment The environment in which the AI agent operates, providing context or resources.
- * @param executor The `PromptExecutor` responsible for executing prompts and managing interactions.
- * @param tools A list of tool descriptors that define the behavior and capabilities of the tools available to the session.
- * @param toolRegistry The registry maintaining a collection of available tools for the session.
- * @param prompt The initial prompt that sets the context or goal for the session.
- * @param model The language model (`LLModel`) used by the session for generating responses and actions.
- * @param responseProcessor An optional processor for handling and transforming model responses, or null if not required.
- * @param config Configuration settings (`AIAgentConfig`) that define session-specific parameters and behavior.
- * @param clock The clock used to track time-related operations within the session.
  */
-public expect class AIAgentLLMWriteSession internal constructor(
+public expect class AIAgentLLMWriteSession
+/**
+ * Internal constructor used by the session infrastructure to create a mutable LLM session.
+ *
+ * @param environment The environment in which the AI agent operates.
+ * @param executor The prompt executor used to send requests to the model.
+ * @param tools A list of tool descriptors available in this session.
+ * @param toolRegistry Registry containing all tools available for resolution by this session.
+ * @param prompt The current prompt used as input for requests.
+ * @param model The active model used to execute requests.
+ * @param responseProcessor Optional response post-processor for model outputs.
+ * @param config Agent configuration used by this session.
+ * @param clock Clock used for timestamped prompt updates and related operations.
+ */
+internal constructor(
     environment: AIAgentEnvironment,
     executor: PromptExecutor,
     tools: List<ToolDescriptor>,
@@ -55,161 +47,7 @@ public expect class AIAgentLLMWriteSession internal constructor(
     responseProcessor: ResponseProcessor?,
     config: AIAgentConfig,
     clock: Clock
-) : AIAgentLLMWriteSessionAPI {
-
-    @PublishedApi
-    internal val delegate: AIAgentLLMWriteSessionImpl
-
-    override val environment: AIAgentEnvironment
-    override val toolRegistry: ToolRegistry
-    override val clock: Clock
-    override var prompt: Prompt
-    override var tools: List<ToolDescriptor>
-    override var model: LLModel
-    override var responseProcessor: ResponseProcessor?
-    override val config: AIAgentConfig
-
-    public override fun <TArgs, TResult> findTool(tool: Tool<TArgs, TResult>): SafeTool<TArgs, TResult>
-
-    public override fun <TArgs, TResult> findTool(toolClass: KClass<out Tool<TArgs, TResult>>): SafeTool<TArgs, TResult>
-
-    public override fun appendPrompt(body: PromptBuilder.() -> Unit)
-
-    @Deprecated("Use `appendPrompt` instead", ReplaceWith("appendPrompt(body)"))
-    public override fun updatePrompt(body: PromptBuilder.() -> Unit)
-
-    public override fun rewritePrompt(body: (prompt: Prompt) -> Prompt)
-
-    public override fun changeModel(newModel: LLModel)
-
-    public override fun changeLLMParams(newParams: LLMParams)
-
-    override suspend fun requestLLMMultipleWithoutTools(): List<Message.Response>
-
-    override suspend fun requestLLMWithoutTools(): Message.Response
-
-    override suspend fun requestLLMOnlyCallingTools(): Message.Response
-
-    override suspend fun requestLLMMultipleOnlyCallingTools(): List<Message.Response>
-
-    override suspend fun requestLLMForceOneTool(tool: ToolDescriptor): Message.Response
-
-    override suspend fun requestLLMForceOneTool(tool: Tool<*, *>): Message.Response
-
-    override suspend fun requestLLM(): Message.Response
-    override suspend fun requestLLMStreaming(): Flow<StreamFrame>
-    override suspend fun requestModeration(moderatingModel: LLModel?): ModerationResult
-
-    override suspend fun requestLLMMultiple(): List<Message.Response>
-
-    override suspend fun <T> requestLLMStructured(
-        config: StructuredRequestConfig<T>,
-        fixingParser: StructureFixingParser?
-    ): Result<StructuredResponse<T>>
-
-    override suspend fun <T> requestLLMStructured(
-        serializer: KSerializer<T>,
-        examples: List<T>,
-        fixingParser: StructureFixingParser?
-    ): Result<StructuredResponse<T>>
-
-    /**
-     * Requests a structured response from the language model session, with optional examples for guidance
-     * and a parser to fix structure-related issues in the response.
-     *
-     * @param T The type of the structured data expected in the response.
-     * @param examples A list of examples to guide the language model in producing the desired structured response.
-     *                 Defaults to an empty list if no examples are provided.
-     * @param fixingParser An optional parser to validate and correct the structure of the language model's response.
-     *                     Defaults to null if no fixing parser is provided.
-     * @return A [Result] containing a [StructuredResponse] of type [T], which includes the structured data or error details.
-     */
-    public suspend inline fun <reified T> requestLLMStructured(
-        examples: List<T> = emptyList(),
-        fixingParser: StructureFixingParser? = null
-    ): Result<StructuredResponse<T>>
-
-    override suspend fun <T> parseResponseToStructuredResponse(
-        response: Message.Assistant,
-        config: StructuredRequestConfig<T>,
-        fixingParser: StructureFixingParser?
-    ): StructuredResponse<T>
-
-    override suspend fun requestLLMMultipleChoices(): List<LLMChoice>
-
-    override fun close()
-
-    public override suspend fun requestLLMStreaming(definition: StructureDefinition?): Flow<StreamFrame>
-
-    /**
-     * Transforms a flow of arguments into a flow of results by asynchronously executing the given tool in parallel.
-     *
-     * @param TArgs the type of the arguments required by the tool.
-     * @param TResult the type of the result produced by the tool, extending ToolResult.
-     * @param safeTool the tool to be executed for each input argument.
-     * @param concurrency the maximum number of parallel executions allowed. Defaults to 16.
-     * @return a flow of results wrapped in SafeTool.Result for each input argument.
-     */
-    public inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCalls(
-        safeTool: SafeTool<TArgs, TResult>,
-        concurrency: Int = 16
-    ): Flow<SafeTool.Result<TResult>>
-
-    /**
-     * Executes a flow of tool arguments in parallel by invoking the provided tool's raw execution method.
-     * Converts each argument in the flow into a string result returned from the tool.
-     *
-     * @param safeTool The tool to execute, wrapped in a SafeTool to ensure safety during execution.
-     * @param concurrency The maximum number of parallel calls to the tool. Default is 16.
-     * @return A flow of string results derived from executing the tool's raw method.
-     */
-    public inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCallsRaw(
-        safeTool: SafeTool<TArgs, TResult>,
-        concurrency: Int = 16
-    ): Flow<String>
-
-    /**
-     * Executes the given tool in parallel for each element in the flow of arguments, up to the specified level of concurrency.
-     *
-     * @param TArgs The type of arguments consumed by the tool.
-     * @param TResult The type of result produced by the tool.
-     * @param tool The tool instance to be executed in parallel.
-     * @param concurrency The maximum number of concurrent executions. Default value is 16.
-     * @return A flow emitting the results of the tool executions wrapped in a SafeTool.Result object.
-     */
-    public inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCalls(
-        tool: Tool<TArgs, TResult>,
-        concurrency: Int = 16
-    ): Flow<SafeTool.Result<TResult>>
-
-    /**
-     * Transforms a Flow of tool argument objects into a Flow of parallel tool execution results, using the specified tool class.
-     *
-     * @param TArgs The type of the tool arguments that the Flow emits.
-     * @param TResult The type of the results produced by the tool.
-     * @param toolClass The class of the tool to be invoked in parallel for processing the arguments.
-     * @param concurrency The maximum number of parallel executions allowed. Default is 16.
-     * @return A Flow containing the results of the tool executions, wrapped in `SafeTool.Result`.
-     */
-    public inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCalls(
-        toolClass: KClass<out Tool<TArgs, TResult>>,
-        concurrency: Int = 16
-    ): Flow<SafeTool.Result<TResult>>
-
-    /**
-     * Converts a flow of arguments into a flow of raw string results by executing the corresponding tool calls in parallel.
-     *
-     * @param TArgs the type of arguments required by the tool.
-     * @param TResult the type of result produced by the tool.
-     * @param toolClass the class of the tool to be invoked.
-     * @param concurrency the number of concurrent tool calls to be executed. Defaults to 16.
-     * @return a flow of raw string results from the parallel tool calls.
-     */
-    public inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCallsRaw(
-        toolClass: KClass<out Tool<TArgs, TResult>>,
-        concurrency: Int = 16
-    ): Flow<String>
-}
+) : AIAgentLLMWriteSessionCommon
 
 /**
  * Executes the specified tool with the given arguments and returns the result within a [SafeTool.Result] wrapper.
