+++ agents/agents-core/src/jvmCommonMain/kotlin/ai/koog/agents/core/agent/session/AIAgentLLMWriteSession.kt
@@ -7,7 +7,6 @@ import ai.koog.agents.annotations.JavaAPI
 import ai.koog.agents.core.agent.config.AIAgentConfig
 import ai.koog.agents.core.annotation.InternalAgentsApi
 import ai.koog.agents.core.environment.AIAgentEnvironment
-import ai.koog.agents.core.environment.SafeTool
 import ai.koog.agents.core.tools.Tool
 import ai.koog.agents.core.tools.ToolDescriptor
 import ai.koog.agents.core.tools.ToolRegistry
@@ -26,29 +25,25 @@ import ai.koog.prompt.structure.StructuredResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.serialization.KSerializer
 import java.util.concurrent.ExecutorService
-import kotlin.reflect.KClass
 import kotlin.time.Clock
 
-@Suppress("DELEGATED_MEMBER_HIDES_SUPERTYPE_OVERRIDE")
-public actual class AIAgentLLMWriteSession internal constructor(
-    @PublishedApi internal actual val delegate: AIAgentLLMWriteSessionImpl
-) : AIAgentLLMWriteSessionAPI by delegate {
-
-    public actual constructor(
-        environment: AIAgentEnvironment,
-        executor: PromptExecutor,
-        tools: List<ToolDescriptor>,
-        toolRegistry: ToolRegistry,
-        prompt: Prompt,
-        model: LLModel,
-        responseProcessor: ResponseProcessor?,
-        config: AIAgentConfig,
-        clock: Clock
-    ) : this(
-        delegate = AIAgentLLMWriteSessionImpl(
-            environment, executor, tools, toolRegistry, prompt, model, responseProcessor, config, clock
-        )
-    )
+/**
+ * JVM actual implementation of a mutable LLM session.
+ *
+ * In addition to common suspend APIs, this class exposes Java-friendly wrappers
+ * that run session operations on the strategy dispatcher.
+ */
+public actual class AIAgentLLMWriteSession actual constructor(
+    environment: AIAgentEnvironment,
+    executor: PromptExecutor,
+    tools: List<ToolDescriptor>,
+    toolRegistry: ToolRegistry,
+    prompt: Prompt,
+    model: LLModel,
+    responseProcessor: ResponseProcessor?,
+    config: AIAgentConfig,
+    clock: Clock
+) : AIAgentLLMWriteSessionCommon(environment, executor, tools, toolRegistry, prompt, model, responseProcessor, config, clock) {
 
     /**
      * Sends a request to the language model without utilizing any tools and returns multiple responses.
@@ -285,34 +280,4 @@ public actual class AIAgentLLMWriteSession internal constructor(
     ): List<LLMChoice> = config.runOnStrategyDispatcher(executorService) {
         requestLLMMultipleChoices()
     }
-
-    public actual inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCalls(
-        safeTool: SafeTool<TArgs, TResult>,
-        concurrency: Int
-    ): Flow<SafeTool.Result<TResult>> = with(delegate) { toParallelToolCallsImpl(safeTool, concurrency) }
-
-    public actual inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCalls(
-        tool: Tool<TArgs, TResult>,
-        concurrency: Int
-    ): Flow<SafeTool.Result<TResult>> = with(delegate) { toParallelToolCallsImpl(tool, concurrency) }
-
-    public actual inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCalls(
-        toolClass: KClass<out Tool<TArgs, TResult>>,
-        concurrency: Int
-    ): Flow<SafeTool.Result<TResult>> = with(delegate) { toParallelToolCallsImpl(toolClass, concurrency) }
-
-    public actual inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCallsRaw(
-        safeTool: SafeTool<TArgs, TResult>,
-        concurrency: Int
-    ): Flow<String> = with(delegate) { toParallelToolCallsRawImpl(safeTool, concurrency) }
-
-    public actual inline fun <reified TArgs, reified TResult> Flow<TArgs>.toParallelToolCallsRaw(
-        toolClass: KClass<out Tool<TArgs, TResult>>,
-        concurrency: Int
-    ): Flow<String> = with(delegate) { toParallelToolCallsRawImpl(toolClass, concurrency) }
-
-    public actual suspend inline fun <reified T> requestLLMStructured(
-        examples: List<T>,
-        fixingParser: StructureFixingParser?
-    ): Result<StructuredResponse<T>> = with(delegate) { requestLLMStructuredImpl(examples, fixingParser) }
 }
